# specs/001-architecture-docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Client/Server Application Platform API
  description: REST API for user management, authentication, RBAC, and Excel import
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /users/{userId}/roles:
    get:
      tags: [Users]
      summary: Get user's assigned roles
      operationId: getUserRoles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User's roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Set user's roles (replaces all existing)
      operationId: setUserRoles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
              required: [role_ids]
      responses:
        '200':
          description: Roles updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /roles:
    get:
      tags: [Roles]
      summary: List all roles
      operationId: listRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Roles]
      summary: Create a new role
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Role name already exists
  /roles/{roleId}:
    parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Roles]
      summary: Get role by ID
      operationId: getRole
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Role details with permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleWithPermissions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Roles]
      summary: Update role
      operationId: updateRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Roles]
      summary: Delete role (cascades user assignments)
      operationId: deleteRole
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Role deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /roles/{roleId}/permissions:
    parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      tags: [Roles]
      summary: Set role permissions (replaces all existing)
      operationId: setRolePermissions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permission_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
              required: [permission_ids]
      responses:
        '200':
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleWithPermissions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /permissions:
    get:
      tags: [Roles]
      summary: List all available permissions
      operationId: listPermissions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Permission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================
  # MODULES
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required: [error, message]

    # Authentication
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
      required: [id, name]
    RoleWithPermissions:
      allOf:
        - $ref: '#/components/schemas/Role'
        - type: object
          properties:
            permissions:
              type: array
              items:
                $ref: '#/components/schemas/Permission'
          required: [permissions]
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        description:
          type: string
        resource:
          type: string
        action:
          type: string
      required: [id, code, description, resource, action]
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
        permission_ids:
          type: array
          items:
            type: string
            format: uuid
      required: [name]
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string

    # Modules
...